Feature/enhacements search by text (#12)
* Adding Flyway migration Search scripts enhacements

GIN Index is required to obtain a better performance in search

* Disable the Search in the Category table.

This search was disable, because the performance was impacted.

* Disable the Search in the Category table for  Search TestIntegration

This search was disable, because the performance was impacted.

* Adding Flyway migration scripts to consider gin index

To obtain a better performance, we add gin index and the scripts create them.

* Adding Hikari configurations to avoid issues

Hikari have some default values, but some others are missing.

* Disable the Cache annotation to execute some test integrations

This include test in the K8 Cluster

* Fix for NOT filter in sections and Disable the Search by Category

* Application properties updates to grant support to CloudFront cdns

Specific cdn urls has been updated with the new values created for CloudFront

* Containers creation  flow updates

Some updates for the Docker file and a docker compose to test before to deploy to K8

* Fix for search in ContentSpecifications