September 10 - 2025 Release Cancun Event (#14)
* Contents database structure updates and enhacements

* Add persistance layer

* rename genero to gender and add data to embebed ids

---------

Co-authored-by: Fer <fer_kail@hotmail.com>

* Refactor DB and add endpoints to get content by id and by parent id

* Add error on content by parent endpoint

* fix infinite recursion on GenreConverter

* review comments, rename and move packages

* Add DB search engine and search endpoint (#4)

* Add DB search engine and search endpoint

* clean up

---------

Co-authored-by: Fer <fer_kail@hotmail.com>

* Clean up and house keeping, small improvements (#5)

Content api with improvements in Exceptions and other artifacts

* Refactor to DB and endpoints to handle multiple categories per content and tag translations search (#6)

Co-authored-by: Fer <fer_kail@hotmail.com>

* Add create content endpoint and add global flag to tags (#7)

Co-authored-by: Fer <fer_kail@hotmail.com>

* Add create content validations and default tags (#8)

Co-authored-by: Fer <fer_kail@hotmail.com>

* Add content sections controller, endgine and refine sections db structure (#9)

* Add content sections controller, endgine and refine sections db structure

* change put by post and fix column name mapping

* change ContentSectionController method names

---------

Co-authored-by: Fer <fer_kail@hotmail.com>

* Fix for Integration Test to config FetchType to EAGER

* Swagger with minimal setup

Minimal configuration but is working as expected

* Postman Collections to retrieve response from the Contents end points

Collection that is very usefull to test the Contents end point

* Profiles enabled and ready to use

Use of secrets and parameter store for devops, staging, and production

* Add number of episodes to contentDto for series

* Add number of episodes to contentDto for series (#11)

Co-authored-by: Fer <fer_kail@hotmail.com>

* Updating Properties files in order to simplify the files

In compliance with PR comment and optimizing the properties

* Adding Docker Improvements to Pre Cluster Deploy

Security improvements

* Improvements in Properties and Env Variables

Adding default  schema currently is missing, and some Swagger fix required

* Update ContentInfoType.java to grant suppor a new attribute

Front End extend a request to Back End to add the EXPERT_COUNTRY_CODE as a new attribute

* Only a clarification to zero results.  Is normal this situation.

* Fix for Integration Test to config FetchType to EAGER

* Swagger with minimal setup

Minimal configuration but is working as expected

* Postman Collections to retrieve response from the Contents end points

Collection that is very usefull to test the Contents end point

* Profiles enabled and ready to use

Use of secrets and parameter store for devops, staging, and production

* Updating Properties files in order to simplify the files

In compliance with PR comment and optimizing the properties

* Adding Docker Improvements to Pre Cluster Deploy

Security improvements

* Improvements in Properties and Env Variables

Adding default  schema currently is missing, and some Swagger fix required

* Update ContentInfoType.java to grant suppor a new attribute

Front End extend a request to Back End to add the EXPERT_COUNTRY_CODE as a new attribute

* Only a clarification to zero results.  Is normal this situation.

* Feature/standardization with profiles and swagger-IS IN INTEGRATION 

* Fix for Integration Test to config FetchType to EAGER

* Swagger with minimal setup

Minimal configuration but is working as expected

* Postman Collections to retrieve response from the Contents end points

Collection that is very usefull to test the Contents end point

* Profiles enabled and ready to use

Use of secrets and parameter store for devops, staging, and production

* Add number of episodes to contentDto for series

* Add number of episodes to contentDto for series (#11)

Co-authored-by: Fer <fer_kail@hotmail.com>

* Updating Properties files in order to simplify the files

In compliance with PR comment and optimizing the properties

* Adding Docker Improvements to Pre Cluster Deploy

Security improvements

* Improvements in Properties and Env Variables

Adding default  schema currently is missing, and some Swagger fix required

* Update ContentInfoType.java to grant suppor a new attribute

Front End extend a request to Back End to add the EXPERT_COUNTRY_CODE as a new attribute

* Only a clarification to zero results.  Is normal this situation.

* Fix for Integration Test to config FetchType to EAGER

* Swagger with minimal setup

Minimal configuration but is working as expected

* Postman Collections to retrieve response from the Contents end points

Collection that is very usefull to test the Contents end point

* Profiles enabled and ready to use

Use of secrets and parameter store for devops, staging, and production

* Updating Properties files in order to simplify the files

In compliance with PR comment and optimizing the properties

* Adding Docker Improvements to Pre Cluster Deploy

Security improvements

* Improvements in Properties and Env Variables

Adding default  schema currently is missing, and some Swagger fix required

* Update ContentInfoType.java to grant suppor a new attribute

Front End extend a request to Back End to add the EXPERT_COUNTRY_CODE as a new attribute

* Only a clarification to zero results.  Is normal this situation.

---------

Co-authored-by: Fer <fer_kail@hotmail.com>
Co-authored-by: fbucm <fbarrera@ultramasterclass.com>

* Feature/enhacements search by text (#12)

* Adding Flyway migration Search scripts enhacements

GIN Index is required to obtain a better performance in search

* Disable the Search in the Category table.

This search was disable, because the performance was impacted.

* Disable the Search in the Category table for  Search TestIntegration

This search was disable, because the performance was impacted.

* Adding Flyway migration scripts to consider gin index

To obtain a better performance, we add gin index and the scripts create them.

* Adding Hikari configurations to avoid issues

Hikari have some default values, but some others are missing.

* Disable the Cache annotation to execute some test integrations

This include test in the K8 Cluster

* Fix for NOT filter in sections and Disable the Search by Category

* Application properties updates to grant support to CloudFront cdns

Specific cdn urls has been updated with the new values created for CloudFront

* Containers creation  flow updates

Some updates for the Docker file and a docker compose to test before to deploy to K8

* Fix for search in ContentSpecifications

* Adding categories in the DISCOVERY and other little enhacenments (#13)

* Adding categories in the DISCOVERY

As Back end team missing requeriment, now we are adding categories into the end point response.

* Fix to grant support for unable the name and content code

Only the tags search is working with good performance

* Manage excludeTags and excludeCategories for Content

With this exclude business rules for Tags and Categories, some videos shouldn't be visible under specific situations

---------

Co-authored-by: fbucm <fbarrera@ultramasterclass.com>
Co-authored-by: Fer <fer_kail@hotmail.com>